refactor: move spec-graph tests to unit test, delete duplicate E2E files
- Delete tests/e2e/ui/stories-draft-preservation.spec.ts (duplicate of stories-drafts.spec.ts)
- Delete tests/e2e/ui/draft-contract.spec.ts (overlaps with stories-drafts.spec.ts)
- Remove 'Spec graph analysis' sections from stories-drafts, stories-sidebar, stories-navigation
- Create tests/spec-graph-analysis.test.ts with all pure-logic spec-graph tests
  (7 tests: defineStory normalization, contractCompleteness, CT-02/CT-03/CT-04/CT-13 coverage, findRelatedStories)
