+++ tests/e2e/ui/stories-navigation.spec.ts
@@ -14,13 +14,7 @@ import { test, expect } from "../gateway-harness.js";
 import { waitForHealth, createSession, deleteSession, createGoal, deleteGoal, apiFetch } from "../e2e-setup.js";
 import {
 	SpecContext,
-	defineStory,
-	defineContract,
-	exportSpecGraph,
-	contractCompleteness,
-	clearStoryRegistry,
 } from "./spec-framework.js";
-import { CT_03, CT_13 } from "./spec-contracts.js";
 import {
 	STORY_N01,
 	STORY_N02,
@@ -513,45 +507,4 @@ test.describe("CT-13: URL routing and navigation", () => {
 	});
 });
 
-// ---------------------------------------------------------------
-// Spec graph analysis for navigation stories
-// ---------------------------------------------------------------
 
-test.describe("Navigation spec graph", () => {
-	test.beforeEach(() => {
-		clearStoryRegistry();
-		// Re-register contract — another spec file's clearContractRegistry() may have wiped it
-		defineContract(CT_13);
-	});
-
-	test("CT-13 coverage from navigation stories", () => {
-		const ct13 = CT_13;
-
-		// Re-define all navigation stories to check coverage
-		defineStory({ id: "N-01", title: "Sidebar session selection", contracts: [ct13], covers: ["view-transitions", "page-reload"] });
-		defineStory({ id: "N-02", title: "Goal dashboard nav and back", contracts: [ct13], covers: ["back-forward-navigation"] });
-		defineStory({ id: "N-03", title: "Deep links to all view types", contracts: [ct13], covers: ["bookmarks"] });
-		defineStory({ id: "N-04", title: "Browser back and forward", contracts: [ct13], covers: ["back-forward-navigation"] });
-		defineStory({ id: "N-06", title: "Sidebar collapse persistence", contracts: [ct13], covers: ["page-reload"] });
-		defineStory({ id: "N-07", title: "Page title", contracts: [ct13], covers: ["page-reload"] });
-		defineStory({ id: "N-08", title: "Keyboard shortcuts", contracts: [ct13], covers: ["view-transitions"] });
-		defineStory({ id: "N-09", title: "Cross-feature journey", contracts: [ct13], covers: ["view-transitions", "back-forward-navigation"] });
-		defineStory({ id: "N-10", title: "Settings sub-navigation", contracts: [ct13], covers: ["view-transitions", "bookmarks"] });
-
-		const results = contractCompleteness();
-		const ct13Report = results.find(r => r.contractId === "CT-13")!;
-
-		expect(ct13Report).toBeTruthy();
-
-		// All 4 CT-13 variations should be covered
-		const covered = ct13Report.variations.filter(v => v.coveredBy !== null);
-		expect(covered).toHaveLength(4);
-
-		// 100% coverage
-		expect(ct13Report.coverage).toBe(1);
-
-		const graph = exportSpecGraph();
-		expect(Object.keys(graph.stories).length).toBe(9);
-		expect(graph.contracts["CT-13"].stories.length).toBe(9);
-	});
-});
